# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bff_project",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "PAYLOAD_TOO_LARGE",
                  "UNPROCESSABLE_ENTITY",
                  "TOO_MANY_REQUESTS",
                  "INTERNAL_ERROR",
                  "BAD_GATEWAY",
                  "SERVICE_UNAVAILABLE",
                  "GATEWAY_TIMEOUT"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/projects/{projectId}/tasks": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Crée une nouvelle tâche pour un projet existant",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Identifiant du projet",
              "example": "project-1"
            },
            "required": true,
            "description": "Identifiant du projet",
            "name": "projectId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "pattern": "^[^<>]*$",
                    "example": "Scan task"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "todo",
                      "in-progress",
                      "review",
                      "done"
                    ],
                    "description": "Le statut du projet",
                    "example": "todo"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "high",
                      "medium",
                      "low"
                    ],
                    "description": "La priorité du projet",
                    "example": "high"
                  },
                  "responsibleId": {
                    "type": "string",
                    "pattern": "^(?:user-\\d+)?$",
                    "example": "user-1"
                  },
                  "assigneeIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^user-\\d+$",
                      "example": "user-1"
                    }
                  },
                  "labels": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[^<>]*$",
                      "example": "security"
                    }
                  },
                  "dueDate": {
                    "type": "string",
                    "example": "2030-12-31T00:00:00Z"
                  }
                },
                "required": [
                  "title",
                  "status",
                  "priority",
                  "responsibleId",
                  "assigneeIds",
                  "labels",
                  "dueDate"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tâche créée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "L'identifiant unique de la tâche",
                      "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "title": {
                      "type": "string",
                      "description": "Le titre de la tâche",
                      "example": "Nouvelle tâche"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "todo",
                        "in-progress",
                        "review",
                        "done"
                      ],
                      "description": "Le statut du projet",
                      "example": "todo"
                    },
                    "statusLabel": {
                      "type": "string",
                      "description": "Le libellé du statut de la tâche",
                      "example": "À faire"
                    },
                    "responsible": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "L'identifiant unique de la personne",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        },
                        "name": {
                          "type": "string",
                          "description": "Le nom de la personne",
                          "example": "John Doe"
                        },
                        "avatarUrl": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "format": "uri",
                          "description": "L'URL de l'avatar de la personne",
                          "example": "https://example.com/avatar.jpg"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "avatarUrl"
                      ],
                      "description": "Une personne assignable à un projet"
                    },
                    "assignees": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "L'identifiant unique de la personne",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                          },
                          "name": {
                            "type": "string",
                            "description": "Le nom de la personne",
                            "example": "John Doe"
                          },
                          "avatarUrl": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "format": "uri",
                            "description": "L'URL de l'avatar de la personne",
                            "example": "https://example.com/avatar.jpg"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "avatarUrl"
                        ],
                        "description": "Une personne assignable à un projet"
                      },
                      "description": "La liste des personnes assignées à la tâche"
                    },
                    "priority": {
                      "type": "string",
                      "enum": [
                        "high",
                        "medium",
                        "low"
                      ],
                      "description": "La priorité du projet",
                      "example": "high"
                    },
                    "priorityLabel": {
                      "type": "string",
                      "description": "Le libellé de la priorité de la tâche",
                      "example": "Moyenne"
                    },
                    "labels": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "La liste des étiquettes associées à la  tâche",
                      "example": [
                        "backend",
                        "important"
                      ]
                    },
                    "dueDate": {
                      "type": "string",
                      "description": "La date d'échéance de la tâche au format ISO 8601",
                      "example": "2024-12-31T23:59:59Z"
                    },
                    "completed": {
                      "type": "boolean",
                      "description": "Indique si la tâche est complétée",
                      "example": false
                    },
                    "createdAt": {
                      "type": "string",
                      "description": "La date de création de la tâche au format ISO 8601",
                      "example": "2024-01-01T12:00:00Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "description": "La date de dernière mise à jour de la tâche au format ISO 8601",
                      "example": "2024-01-15T15:30:00Z"
                    },
                    "permissions": {
                      "type": "object",
                      "properties": {
                        "canView": {
                          "type": "boolean"
                        },
                        "canEdit": {
                          "type": "boolean"
                        },
                        "canDelete": {
                          "type": "boolean"
                        },
                        "canUpdateStatus": {
                          "type": "boolean"
                        },
                        "canComment": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "canView",
                        "canEdit",
                        "canDelete",
                        "canUpdateStatus",
                        "canComment"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "status",
                    "statusLabel",
                    "responsible",
                    "assignees",
                    "priority",
                    "priorityLabel",
                    "labels",
                    "dueDate",
                    "completed",
                    "createdAt",
                    "permissions"
                  ],
                  "description": "Une tâche associée à un projet"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient rights",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "BFF User, Project API or Core API is unreachable or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}